backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.0,<26.0
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0
gunicorn>=22.0,<24.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
